requirements.txt
distro==1.9.0
et-xmlfile==1.1.0
JPype1==1.5.0
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tabula-py==2.9.3
tzdata==2024.1
